+++ rotkehlchen/tests/api/test_exchanges.py
@@ -6,12 +6,13 @@ from urllib.parse import urlencode
 import pytest
 import requests
 
+from rotkehlchen.assets.asset import Asset
 from rotkehlchen.constants.assets import A_BTC
 from rotkehlchen.exchanges.bitfinex import API_KEY_ERROR_MESSAGE as BITFINEX_API_KEY_ERROR_MESSAGE
 from rotkehlchen.exchanges.bitstamp import (
     API_KEY_ERROR_CODE_ACTION as BITSTAMP_API_KEY_ERROR_CODE_ACTION,
 )
-from rotkehlchen.exchanges.data_structures import AssetMovement
+from rotkehlchen.exchanges.data_structures import AssetMovement, Trade
 from rotkehlchen.exchanges.kucoin import API_KEY_ERROR_CODE_ACTION as KUCOIN_API_KEY_ERROR_CODE
 from rotkehlchen.exchanges.manager import SUPPORTED_EXCHANGES
 from rotkehlchen.fval import FVal
@@ -41,7 +42,7 @@ from rotkehlchen.tests.utils.history import (
     prepare_rotki_for_history_processing_test,
 )
 from rotkehlchen.tests.utils.mock import MockResponse
-from rotkehlchen.typing import AssetMovementCategory, Location
+from rotkehlchen.typing import AssetMovementCategory, Location, TradeType
 
 
 def mock_validate_api_key():
@@ -67,6 +68,7 @@ API_KEYPAIR_COINBASE_VALIDATION_PATCH = patch(
 )
 
 
+@pytest.mark.parametrize('number_of_eth_accounts', [0])
 def test_setup_exchange(rotkehlchen_api_server):
     """Test that setting up an exchange via the api works"""
     # Check that no exchanges are registered
@@ -144,6 +146,7 @@ def test_setup_exchange(rotkehlchen_api_server):
     assert json_data['result'] == ['kraken', 'coinbasepro']
 
 
+@pytest.mark.parametrize('number_of_eth_accounts', [0])
 @pytest.mark.parametrize('added_exchanges', [('kraken',)])
 def test_kraken_malformed_response(rotkehlchen_api_server_with_exchanges):
     """Test that if Rotki gets a malformed response from Kraken it's handled properly
@@ -192,6 +195,7 @@ def test_kraken_malformed_response(rotkehlchen_api_server_with_exchanges):
     assert data['message'] == ""
 
 
+@pytest.mark.parametrize('number_of_eth_accounts', [0])
 def test_setup_exchange_does_not_stay_in_mapping_after_500_error(rotkehlchen_api_server):
     """Test that if 500 error is returned during setup of an exchange and it's stuck
     in the exchange mapping Rotki doesn't still think the exchange is registered.
@@ -224,6 +228,7 @@ def test_setup_exchange_does_not_stay_in_mapping_after_500_error(rotkehlchen_api
     assert json_data['result'] == ['kraken']
 
 
+@pytest.mark.parametrize('number_of_eth_accounts', [0])
 def test_setup_exchange_errors(rotkehlchen_api_server):
     """Test errors and edge cases of setup_exchange endpoint"""
 
@@ -312,6 +317,7 @@ def test_setup_exchange_errors(rotkehlchen_api_server):
     )
 
 
+@pytest.mark.parametrize('number_of_eth_accounts', [0])
 def test_remove_exchange(rotkehlchen_api_server):
     """Test that removing a setup exchange via the api works"""
     rotki = rotkehlchen_api_server.rest_api.rotkehlchen
@@ -369,6 +375,7 @@ def test_remove_exchange(rotkehlchen_api_server):
     )
 
 
+@pytest.mark.parametrize('number_of_eth_accounts', [0])
 def test_remove_exchange_errors(rotkehlchen_api_server):
     """Errors and edge cases when using the remove exchange endpoint"""
     # remove unsupported exchange
@@ -439,6 +446,7 @@ def test_exchange_query_balances(rotkehlchen_api_server_with_exchanges):
     assert_poloniex_balances_result(result['poloniex'])
 
 
+@pytest.mark.parametrize('number_of_eth_accounts', [0])
 @pytest.mark.parametrize('added_exchanges', [('binance', 'poloniex')])
 def test_exchange_query_balances_ignore_cache(rotkehlchen_api_server_with_exchanges):
     """Test that using the exchange balances query endpoint can ignore cache"""
@@ -485,6 +493,7 @@ def test_exchange_query_balances_ignore_cache(rotkehlchen_api_server_with_exchan
         assert bn.call_count == 6, 'call count should have changed. Cache should have been ignored'
 
 
+@pytest.mark.parametrize('number_of_eth_accounts', [0])
 @pytest.mark.parametrize('added_exchanges', [('binance', 'poloniex')])
 def test_exchange_query_balances_errors(rotkehlchen_api_server_with_exchanges):
     """Test errors and edge cases of the exchange balances query endpoint"""
@@ -514,6 +523,7 @@ def test_exchange_query_balances_errors(rotkehlchen_api_server_with_exchanges):
     )
 
 
+@pytest.mark.parametrize('number_of_eth_accounts', [0])
 @pytest.mark.parametrize('added_exchanges', [('binance', 'poloniex')])
 def test_exchange_query_trades(rotkehlchen_api_server_with_exchanges):
     """Test that using the exchange trades query endpoint works fine"""
@@ -582,6 +592,7 @@ def test_exchange_query_trades(rotkehlchen_api_server_with_exchanges):
         assert_okay(response)
 
 
+@pytest.mark.parametrize('number_of_eth_accounts', [0])
 @pytest.mark.parametrize('added_exchanges', [('kraken', 'poloniex')])
 def test_query_asset_movements(rotkehlchen_api_server_with_exchanges):
     """Test that using the asset movements query endpoint works fine"""
@@ -670,6 +681,7 @@ def test_query_asset_movements(rotkehlchen_api_server_with_exchanges):
         assert_okay(response)
 
 
+@pytest.mark.parametrize('number_of_eth_accounts', [0])
 @pytest.mark.parametrize('added_exchanges', [('kraken', 'poloniex')])
 @pytest.mark.parametrize('start_with_valid_premium', [False, True])
 def test_query_asset_movements_over_limit(
@@ -736,3 +748,48 @@ def test_query_asset_movements_over_limit(
             assert len(result['entries']) == FREE_ASSET_MOVEMENTS_LIMIT - polo_entries_num
             assert result['entries_limit'] == FREE_ASSET_MOVEMENTS_LIMIT
             assert result['entries_found'] == all_movements_num
+
+
+@pytest.mark.parametrize('number_of_eth_accounts', [0])
+def test_delete_external_exchange_data_works(rotkehlchen_api_server_with_exchanges):
+    server = rotkehlchen_api_server_with_exchanges
+    rotki = server.rest_api.rotkehlchen
+
+    trades = [Trade(
+        timestamp=0,
+        location=x,
+        pair='ETH_EUR',
+        trade_type=TradeType.BUY,
+        amount=FVal(1),
+        rate=FVal(1),
+        fee=FVal(1),
+        fee_currency=Asset('EUR'),
+        link='',
+        notes='',
+    ) for x in (Location.CRYPTOCOM, Location.KRAKEN)]
+    rotki.data.db.add_trades(trades)
+    movements = [AssetMovement(
+        location=x,
+        category=AssetMovementCategory.DEPOSIT,
+        address=None,
+        transaction_id=None,
+        timestamp=0,
+        asset=A_BTC,
+        amount=FVal(100),
+        fee_asset=A_BTC,
+        fee=FVal(1),
+        link='') for x in (Location.CRYPTOCOM, Location.KRAKEN)]
+    rotki.data.db.add_asset_movements(movements)
+    assert len(rotki.data.db.get_trades()) == 2
+    assert len(rotki.data.db.get_asset_movements()) == 2
+    response = requests.delete(
+        api_url_for(
+            server,
+            'named_exchanges_data_resource',
+            name='crypto.com',
+        ),
+    )
+    result = assert_proper_response_with_result(response)  # just check no validation error happens
+    assert result is True
+    assert len(rotki.data.db.get_trades()) == 1
+    assert len(rotki.data.db.get_asset_movements()) == 1